Moves credit type to the credtrack and credmod packages
