+++ packages/credtrack/credmod/currency.go
@@ -1,0 +1,7 @@
+package credmod
+
+type Currency string
+
+const (
+	GBP Currency = "GBP"
+)
