+++ packages/credtrack/credmod/monetary_amount.go
@@ -1,0 +1,6 @@
+package credmod
+
+type MonetaryAmount struct {
+	Value    float32  `dynamodbav:value`
+	Currency Currency `dynamodbav:currency`
+}
