+++ packages/credtrack/credmod/transaction_type.go
@@ -1,0 +1,8 @@
+package credmod
+
+type TransactionType string
+
+const (
+	CREDIT TransactionType = "CREDIT"
+	DEBIT                  = "DEBIT"
+)
