+++ packages/credtrack/go.mod
@@ -1,0 +1,3 @@
+module theodo.red/creditcompanion/packages/credtrack
+
+go 1.15
